Now fetches profile data and stores into database
This change allows the data (profile name, and profile image URL) to be stored in the database. Doing this method allows the review page to simply grab the url instead of making multiple API calls on load to grab the data. Using this we can have all the data preloaded with no hovering speeding up the process of reviewing immensly